requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
tensorflow==2.15.0
mysql-connector-python==8.3.0
pydantic==2.7.1
werkzeug==3.0.1